Add a fair amount of already working things
This sets up and is a functional FF extension with the following:
 * Background Page for long running processes/state
 * Options UI for settings and preferences etc
 * Popup with a Page Action button

@emotion for CSS-in-JS, Tailwind CSS because I don't like writing a lot of CSS
Storybook for viewing and debugging components, along with some additional plugins
Immer for state management
(haven't brought in my typed redux but with-hooks-and-immer clone yet)

Gets me ready to start re-implementing the current sandboxed 2.0 web
extension which I originally started with an alpha release of hooks and
semantic-ui. The hope being to take lessons learned and implement
a better and more stable (and ready for public release) extension finally.